feat(vs2022_third): C4018 fix signed/unsigned mismatch in NanoSpec — 6 files (batch 3)
Cast int loop variables to size_t/DWORD in comparisons:
- NanoSpec/RecipeStressStageProgramView.cpp: 8 instances (i,j→DWORD/unsigned int)
- NanoSpec/RecipeThicknessSettingView.cpp: 6 instances (iCount→DWORD)
- NanoSpec/RecipeMainSrView.cpp: 3 instances (i,j,iCount→DWORD)
- NanoSpec/ChifTransiStress.cpp: 2 instances (i→DWORD)
- MeaSys/Xmp.cpp: 1 instance (i→size_t)

Total: 20 C4018 warnings fixed.
No behavior changes, only explicit casts to resolve type mismatch.

+++ SRC/NanoSpec/ChifTransiStress.cpp
@@ -937,7 +937,7 @@ void CChiefTransiStress::enter_ST_STRS_LMVSTAPOS(const cEventParams* const pEvPa
 //																			/*	セクション１の Y 座標を使います．	*/
 ///* modified 2009.08.18 hmenjo STRESS_LINESECTION 削除 ---------- } ---------- */
 /* modified 2009.08.21 hmenjo ストレス機能追加(51) ----------			   */
-			for (int i = 0; i < l_StressConfig.Line[m_dwLineNo - 1].dwSectionNum; i++) {
+			for (int i = 0; (DWORD)i < l_StressConfig.Line[m_dwLineNo - 1].dwSectionNum; i++) {
 				if (0 != l_pStageProgStress->Line[m_dwLineNo - 1].bScanValid[i]) {
 					break;
 				}
@@ -1125,7 +1125,7 @@ void CChiefTransiStress::enter_ST_STRS_LMVENDPOS(const cEventParams* const pEvPa
 //																					/*	セクション１の Y 座標を使います．	*/
 ///* modified 2009.08.18 hmenjo STRESS_LINESECTION 削除 ---------- } ---------- */
 /* modified 2009.08.21 hmenjo ストレス機能追加(51) ----------			   */
-					for (int i = 0; i < l_StressConfig.Line[m_dwLineNo - 1].dwSectionNum; i++) {
+					for (int i = 0; (DWORD)i < l_StressConfig.Line[m_dwLineNo - 1].dwSectionNum; i++) {
 						if (0 != l_pStageProgStress->Line[m_dwLineNo - 1].bScanValid[i]) {
 							break;
 						}
