+++ SRC/NanoSpec/RecipeStressStageProgramView.cpp
@@ -982,7 +982,7 @@ void CRecipeStressStageProgramView::SetMeasurementLineSectionData()
 			continue;
 		}
 
-		for (j=0; j<m_StressConfig.Line[i].dwSectionNum; j++) {
+		for (j=0; (DWORD)j<m_StressConfig.Line[i].dwSectionNum; j++) {
 			dwTextStyle = DT_CENTER|DT_VCENTER|DT_SINGLELINE;
 			dwTextStyle |= DT_END_ELLIPSIS;
 			Item.mask = GVIF_TEXT|GVIF_FORMAT;
@@ -1075,7 +1075,7 @@ void CRecipeStressStageProgramView::SetMeasurementPointData()
 		//2009.09.19 bagus stress --}--
 
 		bUseLine[i] = FALSE;
-		for (j=0; j<m_StressConfig.Line[i].dwSectionNum; j++) {
+		for (j=0; (DWORD)j<m_StressConfig.Line[i].dwSectionNum; j++) {
 			//2009.09.19 bagus stress --{--
 			if (m_StageStressProg.Line[i].bScanValid[j]) {
 			//2009.09.19 bagus stress --}--
@@ -1457,7 +1457,7 @@ void CRecipeStressStageProgramView::UpDate()
 		//2009.09.19 bagus stress --}--
 			continue;
 		}
-		for (int j=0; j<m_StressConfig.Line[n].dwSectionNum; j++) {
+		for (int j=0; (DWORD)j<m_StressConfig.Line[n].dwSectionNum; j++) {
 			//if(m_LineCheckBox.CheckBox[ls_rows-1]->GetCheck()){
 			//2009.09.15 bagus stress --{--
 			//if(((CGridCellCheck *)m_MeasurementPointGrid.GetCell(ls_rows-1,4))->GetCheck()){
@@ -1817,14 +1817,14 @@ void CRecipeStressStageProgramView::SamplePointGraph_DataSet()
 // 2009.09.17 bagus stress GridèCê≥ --}--
 
 	ls_rows = 1;
-	for (i=0; i<m_StressConfig.dwLiftPinNumberOfLine; i++) {
+	for (i=0; (DWORD)i<m_StressConfig.dwLiftPinNumberOfLine; i++) {
 		if(i >= STRESS_PINS_MAX){
 			break;
 		}
 
 //2009.09.10 bagus stress --{--
 // Pin Data
-		for(j = 0; j<m_StressConfig.Line[i].dwLiftPinNumberOfLn;j++){
+		for(j = 0; (DWORD)j<m_StressConfig.Line[i].dwLiftPinNumberOfLn;j++){
 			long lPinX,lPinY;
 			lPinX = lPinY = 0L;
 			switch (m_StageConfig.Dir.X) {
@@ -1861,7 +1861,7 @@ void CRecipeStressStageProgramView::SamplePointGraph_DataSet()
 			m_SamplePointGraph.AddPinData(lPinX,lPinY);
 		}
 //2009.09.10 bagus stress --}--
-		for (j=0; j<m_StressConfig.Line[i].dwSectionNum; j++) {
+		for (j=0; (DWORD)j<m_StressConfig.Line[i].dwSectionNum; j++) {
 // 2009.09.10 bagus stress èCê≥ --{--
 //			if (m_StageStressProg.Line[i].bValidLine) {
 //				tLineData.Status = CNanoPointGraph::SECTION_NOT_MEASURED;
@@ -2560,7 +2560,7 @@ BOOL CRecipeStressStageProgramView::CheckMeasurementLineSectionData()
 // 2009.09.10 bagus stress èCê≥ --{--
 		bUseLine = FALSE;
 // 2009.09.10 bagus stress èCê≥ --}--
-		for (j=0; j<m_StressConfig.Line[i].dwSectionNum; j++) {
+		for (j=0; (DWORD)j<m_StressConfig.Line[i].dwSectionNum; j++) {
 			Item.mask = GVIF_TEXT;
 			Item.row = ls_rows;
 			Item.col = 4;
@@ -2738,7 +2738,7 @@ void CRecipeStressStageProgramView::InitCombo_MaterialElasticModulus()
 	CComboBox* pCombo;
 	pCombo = (CComboBox*)GetDlgItem(IDC_MATERIAL_ELASTIC_MODULUS);
 	pCombo->ResetContent();
-	for (i=0; i<m_ElasticTable.uiTableNum; i++) {
+	for (i=0; (unsigned int)i<m_ElasticTable.uiTableNum; i++) {
 		strItem.Format("%e %s", m_ElasticTable.ElasticModulus[i].dElasticModulusValue, m_ElasticTable.ElasticModulus[i].tszElasticModulusName);
 		pCombo->AddString(strItem);
 	}
